requirements.txt
flask==2.0.1
websocket-client==1.3.3
gunicorn==20.1.0